work on advquery
