+++ advquery.js
@@ -1,25 +1,237 @@
 var db = require('./db');
 var display = require('./display');
 var computeMatchData = display.computeMatchData;
-var filter = display.filter;
-var aggregator = display.aggregator;
 var utility = require('./utility');
 var isRadiant = utility.isRadiant;
-module.exports = function advQuery(options, cb) {
+var mergeObjects = utility.mergeObjects;
+var constants = require('./constants.json');
+
+function aggregator(matches, fields) {
+    var types = {
+        "start_time": function(key, m, p) {
+            agg(key, m.start_time, m);
+        },
+        "duration": function(key, m, p) {
+            agg(key, m.duration, m);
+        },
+        "cluster": function(key, m, p) {
+            agg(key, m.cluster, m);
+        },
+        "first_blood_time": function(key, m, p) {
+            agg(key, m.first_blood_time, m);
+        },
+        "lobby_type": function(key, m, p) {
+            agg(key, m.lobby_type, m);
+        },
+        "game_mode": function(key, m, p) {
+            agg(key, m.game_mode, m);
+        },
+        "hero_id": function(key, m, p) {
+            agg(key, p.hero_id, m);
+        },
+        "kills": function(key, m, p) {
+            agg(key, p.kills, m);
+        },
+        "deaths": function(key, m, p) {
+            agg(key, p.deaths, m);
+        },
+        "assists": function(key, m, p) {
+            agg(key, p.assists, m);
+        },
+        "last_hits": function(key, m, p) {
+            agg(key, p.last_hits, m);
+        },
+        "denies": function(key, m, p) {
+            agg(key, p.denies, m);
+        },
+        "gold_per_min": function(key, m, p) {
+            agg(key, p.gold_per_min, m);
+        },
+        "xp_per_min": function(key, m, p) {
+            agg(key, p.xp_per_min, m);
+        },
+        "hero_damage": function(key, m, p) {
+            agg(key, p.hero_damage, m);
+        },
+        "tower_damage": function(key, m, p) {
+            agg(key, p.tower_damage, m);
+        },
+        "hero_healing": function(key, m, p) {
+            agg(key, p.hero_healing, m);
+        },
+        "leaver_status": function(key, m, p) {
+            agg(key, p.leaver_status, m);
+        },
+        "isRadiant": function(key, m, p) {
+            agg(key, isRadiant(p), m);
+        },
+        "stuns": function(key, m, p) {
+            agg(key, p.parsedPlayer.stuns, m);
+        },
+        "lane": function(key, m, p) {
+            agg(key, p.parsedPlayer.lane, m);
+        },
+        "lane_role": function(key, m, p) {
+            agg(key, p.parsedPlayer.lane_role, m);
+        },
+        //lifetime ward positions
+        "obs": function(key, m, p) {
+            agg(key, p.parsedPlayer.obs, m);
+        },
+        "sen": function(key, m, p) {
+            agg(key, p.parsedPlayer.sen, m);
+        },
+        //lifetime rune counts
+        "runes": function(key, m, p) {
+            agg(key, p.parsedPlayer.runes, m);
+        },
+        //lifetime item uses
+        "item_uses": function(key, m, p) {
+            agg(key, p.parsedPlayer.item_uses, m);
+        },
+        //track sum of purchase times and counts to get average build time
+        "purchase_time": function(key, m, p) {
+            agg(key, p.parsedPlayer.purchase_time, m);
+        },
+        "purchase_time_count": function(key, m, p) {
+            agg(key, p.parsedPlayer.purchase_time_count, m);
+        },
+        "purchase": function(key, m, p) {
+            agg(key, p.parsedPlayer.purchase, m);
+        },
+        "kills_count": function(key, m, p) {
+            agg(key, p.parsedPlayer.kills, m);
+        },
+        "gold_reasons": function(key, m, p) {
+            agg(key, p.parsedPlayer.gold_reasons, m);
+        },
+        "xp_reasons": function(key, m, p) {
+            agg(key, p.parsedPlayer.xp_reasons, m);
+        },
+        "ability_uses": function(key, m, p) {
+            agg(key, p.parsedPlayer.ability_uses, m);
+        },
+        "hero_hits": function(key, m, p) {
+            agg(key, p.parsedPlayer.hero_hits, m);
+        },
+        "courier_kills": function(key, m, p) {
+            agg(key, p.parsedPlayer.courier_kills, m);
+        },
+        "tower_kills": function(key, m, p) {
+            agg(key, p.parsedPlayer.tower_kills, m);
+        },
+        "neutral_kills": function(key, m, p) {
+            agg(key, p.parsedPlayer.neutral_kills, m);
+        },
+        "chat_message_count": function(key, m, p) {
+            agg(key, p.parsedPlayer.chat_message_count, m);
+        },
+        "gg_count": function(key, m, p) {
+            agg(key, p.parsedPlayer.gg_count, m);
+        },
+        "buyback_count": function(key, m, p) {
+            agg(key, p.parsedPlayer.buyback_count, m);
+        },
+        "observer_uses": function(key, m, p) {
+            agg(key, p.parsedPlayer.observer_uses, m);
+        },
+        "sentry_uses": function(key, m, p) {
+            agg(key, p.parsedPlayer.sentry_uses, m);
+        }
+    };
+    var aggData = {};
+    fields = fields || types;
+    for (var type in fields) {
+        aggData[type] = {
+            sum: 0,
+            min: Number.MAX_VALUE,
+            max: 0,
+            max_match: null,
+            n: 0,
+            counts: {},
+        };
+    }
+    for (var i = 0; i < matches.length; i++) {
+        var m = matches[i];
+        var p = m.players[0];
+        for (var type in fields) {
+            if (types[type]) {
+                types[type](type, m, p);
+            }
+        }
+    }
+    return aggData;
+
+    function agg(key, value, match) {
+        var m = aggData[key];
+        if (typeof value === "undefined") {
+            return;
+        }
+        m.n += 1;
+        if (typeof value === "object") {
+            mergeObjects(m.counts, value);
+        }
+        else {
+            if (!m.counts[value]) {
+                m.counts[value] = 0;
+            }
+            m.counts[value] += 1;
+            m.sum += (value || 0);
+            if (value < m.min) {
+                m.min = value;
+            }
+            if (value > m.max) {
+                m.max = value;
+                m.max_match = match;
+            }
+        }
+    }
+}
+
+function filter(matches, filters) {
+    console.log(filters);
+    //accept a hash of filters, run all the filters in the hash in series
+    var filtered = [];
+    for (var key in filters) {
+        for (var i = 0; i < matches.length; i++) {
+            if (key === "balanced") {
+                if (constants.modes[matches[i].game_mode].balanced && constants.lobbies[matches[i].lobby_type].balanced) {
+                    filtered.push(matches[i]);
+                }
+            }
+            else if (key === "win") {
+                if (isRadiant(matches[i].players[0]) === matches[i].radiant_win) {
+                    filtered.push(matches[i]);
+                }
+            }
+            else if (key === "hero_id") {
+                if (matches[i].players[0].hero_id === Number(filters["hero_id"])) {
+                    filtered.push(matches[i]);
+                }
+            }
+            else {
+                filtered.push(matches[i]);
+            }
+        }
+        matches = filtered.slice(0);
+        filtered = [];
+    }
+    return matches;
+}
+
+function advQuery(options, cb) {
     //options passed:
     //select,the query to send to mongodb
     //project
     //filter,additional post-processing filters
+    //agg, aggregations to do
     //limit
     //skip
     //sort
-    //todo implement more features
-    //api caps limit to prevent bandwidth abuse
-    //no cap if calling internally
     //matches page, want matches fitting criteria
     //player matches page, want winrate, matches fitting criteria, also need to display player information
-    //player trends page, want agregation on matches fitting criteria
-    //client options should include:
+    //player trends page, want aggregation on matches fitting criteria
+    //todo implement more features
     //filter: specific player/specific hero id
     //filter: specific player was also in the game (use players.account_id with $and, but which player gets returned by projection?  or could just do in js)
     //filter: specific hero was played by me, on my team, was against me, was in the game
@@ -28,25 +240,37 @@ module.exports = function advQuery(options, cb) {
     //filter: specific regions
     //filter: no stats recorded (need to implement filter to detect)
     //filter: significant game modes only (balanced filter)
-    //CAREFUL! during aggregation not all fields may exist (since we projected)
+    //filter: win
+    //filter kill differential, gold/xp differential?
     //limit/skip are useful for datatables server-side ajax calls, but also prevent aggregation from working properly
     //aggData contains games/win/lose, useful for reporting winrate given a query
-    //The problem with using datatables to interface with the api is that the query string suddenly gets a lot harder to build. 
-    //do we even need a "general" matches page anymore?
-    //Maybe we can add helpers to construct the query, use same advquery UI for:
-    //matches tab (lists matches, reports winrate fitting the advanced query conditions)
-    //trends tab (aggregates data fitting the advanced query conditions)
-    if (options.select["players.account_id"]) {
-        //convert the passed account id to number
-        //todo run this on api end?
-        options.select["players.account_id"] = Number(options.select["players.account_id"]);
-    }
+    //todo have this function return unfiltered data as well?  if so, we dont want to display it in api
+    options.project = options.project || {
+        start_time: 1,
+        match_id: 1,
+        cluster: 1,
+        game_mode: 1,
+        duration: 1,
+        radiant_win: 1,
+        parse_status: 1,
+        first_blood_time: 1,
+        lobby_type: 1
+    };
+    console.log(options.select);
     if (options.select["players.account_id"] || options.select["players.hero_id"]) {
         //if selecting by account_id or hero_id, we project only that user in players array
-        //todo add more cases where there's a single user to aggregate on
-        //todo better detection condition, what if elemmatch syntax?
         options.project["players.$"] = 1;
     }
+    else {
+        //otherwise, always project a player to prevent computematchdata crash
+        options.project.players = {
+            $slice: 1
+        };
+        options.project["players.account_id"] = 1;
+    }
+    //cap the number of matches to analyze
+    var max = 20000;
+    options.limit = (!options.limit || options.limit > max) ? max : options.limit;
     //build the monk hash
     var monk = {
         limit: options.limit,
@@ -70,7 +294,7 @@ module.exports = function advQuery(options, cb) {
         var filtered = filter(matches, options.filter);
         //console.timeEnd('filter');
         //console.time('agg');
-        var aggData = aggregator(filtered);
+        var aggData = aggregator(filtered, options.agg);
         //console.timeEnd('agg');
         //console.time('post');
         aggData.win = 0;
@@ -91,4 +315,5 @@ module.exports = function advQuery(options, cb) {
         };
         cb(err, result);
     });
-};
+}
+module.exports = advQuery;
