+++ sources.json
@@ -134,7 +134,7 @@
     },
     "4": {
       "id": 4,
-      "name": "Co-Op with bots"
+      "name": "Co-Op with Bots"
     },
     "5": {
       "id": 5,
